Refactor extract src & macros folders
